Resolved circular dependency
